Fix melee render squashing
This is done by rendering the textures independently of the size of the
body (scaled to World Units) and repositioning them so that they align
with the center of the body.

Note that it was necessary to add "spacing" on the side of the melee
sprite opposite the sword so that it would be centered in the same
place.

The normals still look a little bit out of wack; @abrinkman94 can you
verify and correct?